fix bugs when creating VCF headers for tebreak and temp2

+++ scripts/tebreak/tebreak_post.py
@@ -72,11 +72,11 @@ def read_insertions(tebreak_out, sample_name, chromosomes, config):
                 insert.support_info.support["three_p_elt_match"].value = float(split_line[header['3p_Elt_Match']])
                 insert.support_info.support["five_p_genome_match"].value = float(split_line[header['5p_Genome_Match']])
                 insert.support_info.support["three_p_genome_match"].value = float(split_line[header['3p_Genome_Match']])
-                insert.support_info.support["split_reads_5prime"].value = float(split_line[header['Split_reads_5prime']])
-                insert.support_info.support["split_reads_3prime"].value = float(split_line[header['Split_reads_3prime']])
-                insert.support_info.support["remapped_discordant"].value = float(split_line[header['Remapped_Discordant']])
+                insert.support_info.support["split_reads_5prime"].value = int(split_line[header['Split_reads_5prime']])
+                insert.support_info.support["split_reads_3prime"].value = int(split_line[header['Split_reads_3prime']])
+                insert.support_info.support["remapped_discordant"].value = int(split_line[header['Remapped_Discordant']])
                 insert.support_info.support["remap_disc_fraction"].value = float(split_line[header['Remap_Disc_Fraction']])
-                insert.support_info.support["remapped_splitreads"].value = float(split_line[header['Remapped_Splitreads']])
+                insert.support_info.support["remapped_splitreads"].value = int(split_line[header['Remapped_Splitreads']])
                 insert.support_info.support["remap_split_fraction"].value = float(split_line[header['Remap_Split_Fraction']])
 
                 insert.name = insert.family+"|non-reference|NA|"+sample_name+"|tebreak|sr|"
